pyproject.toml
[project]
name = "budget-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "fastapi[standard]>=0.116.1",
    "isort>=6.0.1",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.3",
    "langgraph-supervisor>=0.0.29",
    "langsmith>=0.4.12",
    "python-dotenv>=1.1.1",
]

[tool.black]
line-length = 89
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 89
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
